+++ packages/catalog-model/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/catalog-model
 
+## 1.7.8-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/errors@1.3.0-next.0
+
+## 1.7.7
+
+### Patch Changes
+
+- a49a40d: Updated dependency `zod` to `^3.25.76 || ^4.0.0` & migrated to `/v3` or `/v4` imports.
+
 ## 1.7.6
 
 ### Patch Changes